userservice/requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
